refactor code to add StretchHelpers.py
