+++ nodes/StretchSkill.py
@@ -45,6 +45,9 @@ import torch
 from dl2_lfd.helper_funcs.conversions import np_to_pgpu
 # from gazebo_ros_link_attacher.srv import Attach, AttachRequest, AttachResponse
 
+
+from StretchHelpers import feedbackLin, thresholdVel, findCommands, findArmExtensionAndRotation, findTheta
+
 DEVICE="cpu"
 
 # import stretch_funmap.navigate as nv
@@ -313,59 +316,6 @@ class StretchSkill(hm.HelloNode):
 
         return True
 
-def feedbackLin(arg_cmd_vx, arg_cmd_vy, arg_theta, arg_epsilon):
-    cmd_vi = np.array([[arg_cmd_vx], [arg_cmd_vy]])
-    R_b_i = np.array([[np.cos(arg_theta), np.sin(arg_theta)], [-np.sin(arg_theta), np.cos(arg_theta)]])
-    cmd_vw = np.dot(np.dot(np.array([[1, 0], [0, 1/arg_epsilon]]), R_b_i), cmd_vi)
-    cmd_v = cmd_vw[0]
-    cmd_w = cmd_vw[1]
-
-    return cmd_v, cmd_w
-
-def thresholdVel(arg_cmd_v, arg_cmd_w, arg_maxV, arg_wheel2center):
-    maxW = arg_maxV/arg_wheel2center
-    ratioV = np.abs(arg_cmd_v/arg_maxV)
-    ratioW = np.abs(arg_cmd_w/maxW)
-    ratioTot = ratioV + ratioW
-    if ratioTot > 1:
-        arg_cmd_v = arg_cmd_v/ratioTot
-        arg_cmd_w = arg_cmd_w/ratioTot
-
-    return arg_cmd_v, arg_cmd_w
-
-def findCommands(arg_cur_pose, arg_desired_pose):
-    cmd_vx = arg_desired_pose[0] - arg_cur_pose.translation.x
-    cmd_vy = arg_desired_pose[1] - arg_cur_pose.translation.y
-    q1 = arg_cur_pose.rotation.x
-    q2 = arg_cur_pose.rotation.y
-    q3 = arg_cur_pose.rotation.z
-    q0 = arg_cur_pose.rotation.w
-    theta = np.arctan2(2 * (q1 * q2 + q0 * q3), q0 ** 2 + q1 ** 2 - q2 ** 2 - q3 **2)
-    if not IS_SIM:
-        theta -= np.pi
-    return cmd_vx, cmd_vy, theta
-
-def findTheta(arg_cur_pose):
-    """Finds the orientation of the robot given the pose with quaternion info
-    Given the pose of the robot as a transfrom, returns the orientation aka
-    theta of the robot. Uses the formula here:
-    https://en.wikipedia.org/wiki/Conversion_between_quaternions_and_Euler_angles
-    Args:
-        arg_cur_pose: transform_msg
-    Returns:
-        theta: np.array
-    """
-    q1 = arg_cur_pose.rotation.x
-    q2 = arg_cur_pose.rotation.y
-    q3 = arg_cur_pose.rotation.z
-    q0 = arg_cur_pose.rotation.w
-    # theta = np.arctan2(2 * (q1 * q2 + q0 * q3), q0 ** 2 + q1 ** 2 - q2 ** 2 - q3 **2)
-    theta = np.arctan2(2 * (q0 * q3 + q1 * q2), 1 - 2 * (q2 * q2 + q3 * q3))
-    # if not IS_SIM:
-        # theta -= np.pi
-    return theta
-
-
 def findTrajectoryFromDMP(start_pose, end_pose, skill_name, dmp_folder, opts):
     model = DMPNN(opts['start_dimension'], 1024, opts['dimension'], opts['basis_fs']).to(DEVICE)
     model.load_state_dict(torch.load(dmp_folder + skill_name + ".pt"))
@@ -414,66 +364,6 @@ def addJointValuesToPose(arg_stretch_pose, arg_joint_values):
     arg_stretch_pose[5] = arg_joint_values[2]
 
     return arg_stretch_pose
-
-
-def findArmExtensionAndRotation(goal_pose, robot_pose):
-    GtoW = 0.23 #Grip to Wrist Distance
-
-    xd = goal_pose.translation.x
-    yd = goal_pose.translation.y
-
-    xr = robot_pose.translation.x
-    yr = robot_pose.translation.y
-
-    qrobot = findTheta(robot_pose)
-    print("ThetaRobot -> ", qrobot)
-    qr = qrobot - (np.pi)/2
-    print("thetaARM -> ",qr)
-
-    p = yr - (xr*np.tan(qr)) - yd
-
-    #Solving quadratic equation
-
-    a = 1 + (np.tan(qr)**2)
-    b = (-2*xd) + (2*p*np.tan(qr))
-    c = xd**2 + p**2 - GtoW**2
-
-    # Discriminant
-
-    d = (b**2) - (4*a*c)
-    # X values
-
-    x1 = (-b + np.sqrt(d))/(2*a)
-    x2 = (-b - np.sqrt(d))/(2*a)
-
-    y1 = (np.tan(qr) * x1) + yr - (xr*np.tan(qr))
-    y2 = (np.tan(qr) * x2) + yr - (xr*np.tan(qr))
-
-    pt1 = (x1,y1)
-    pt2 = (x2,y2)
-
-    robot = (xr,yr)
-
-    dist1 = dist(pt1,robot)
-    dist2 = dist(pt2,robot)
-
-    if dist1<dist2:
-        amount_to_extend = dist1
-
-    else:
-         amount_to_extend = dist2
-
-    print("AmountEXTEND -> ", amount_to_extend)
-    g = (yd - yr - (amount_to_extend * np.sin(qr)))/GtoW
-    print("Value of g -> ", g)
-    print("sin inverse ->",np.arcsin(g))
-    wrist_theta = np.arcsin(g) - qr
-    print("ThetaWrist -> ", wrist_theta)
-
-    # amount_to_extend = 0
-    # wrist_theta = 0
-
-    return amount_to_extend, wrist_theta
 
 
 if __name__ == '__main__':
